requirements.txt
python-dotenv
fastapi[all]
pyarrow
requests
aiohttp
inewave
idecomp
pandas
pybase62